Modulariesierung nochmal verbessert
Funktionen kompackter geschrieben und weiter ausgelagert
